Switch to protobuf structs

+++ scanner/scanners_test.go
@@ -10,7 +10,7 @@ func TestScanBuiltIn(t *testing.T) {
 	ports, err := GoScan("google.com", []int{80})
 	assert.Nil(t, err)
 	assert.Len(t, ports, 1)
-	assert.Equal(t, ports[0].Port, 80)
+	assert.Equal(t, ports[0].Port, int32(80))
 	assert.True(t, ports[0].Open)
 }
